backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
PyPDF2==3.0.1
python-docx==1.1.2
google-generativeai==0.8.3
python-dotenv==1.0.1
